refactor: extract redux reducer to its own file
